+++ src/component/test.vue
@@ -1,0 +1,11 @@
+<script>
+    export default {
+        render(h){
+            return h('div', [
+                this.$scopedSlots.default({
+                    index: 1
+                })
+            ]);
+        }
+    };
+</script>