+++ src/js/app/main/index.js
@@ -8,7 +8,7 @@ import Buttons from 'common/buttons.vue';
 import RtInput from 'common/rt-input.vue';
 import utility from 'ct-utility';
 import loading from 'ct-adc-loading';
-
+import pagination from 'ct-adc-page';
 Vue.component('adc-table-operator', Buttons);
 Vue.component('adc-img', Img);
 Vue.component('rt-input', RtInput);
@@ -16,12 +16,13 @@ Vue.use(loading);
 Vue.filter('a', function(v) {
     return v * 2;
 });
-Vue.filter('dateTimeFormat', function(v){
+Vue.filter('dateTimeFormat', function(v) {
     return utility.dateFilter(v, 'yyyy-MM-dd HH:mm:ss');
 });
 
 const tableData = [
     {
+        id: 1,
         name: 'ruby',
         age: 12,
         dep: 1,
@@ -29,7 +30,9 @@ const tableData = [
         html: 'hello',
         time1: +new Date() + 86400000,
         time2: +new Date()
-    }, {
+    },
+    {
+        id: 2,
         name: 'ruby2',
         age: 15,
         dep: 1,
@@ -37,7 +40,9 @@ const tableData = [
         html: 'hello',
         time1: +new Date() + 86400000,
         time2: +new Date()
-    }, {
+    },
+    {
+        id: 3,
         name: 'ruby',
         age: 12,
         dep: 1,
@@ -45,7 +50,9 @@ const tableData = [
         html: 'hello',
         time1: +new Date() + 86400000,
         time2: +new Date()
-    }, {
+    },
+    {
+        id: 4,
         name: 'ruby2',
         age: 16,
         dep: 1,
@@ -53,7 +60,11 @@ const tableData = [
         html: 'hello',
         time1: +new Date() + 86400000,
         time2: +new Date()
-    }, {
+    }
+];
+var tableData2 = [
+    {
+        id: 5,
         name: 'ruby',
         dep: 3,
         age: 12,
@@ -61,7 +72,9 @@ const tableData = [
         html: 'hello',
         time1: +new Date() + 86400000,
         time2: +new Date()
-    }, {
+    },
+    {
+        id: 6,
         name: 'ruby2',
         age: 15,
         dep: 1,
@@ -69,7 +82,9 @@ const tableData = [
         html: 'hello',
         time1: +new Date() + 86400000,
         time2: +new Date()
-    }, {
+    },
+    {
+        id: 7,
         name: 'ruby',
         age: 12,
         dep: 2,
@@ -77,7 +92,9 @@ const tableData = [
         html: 'hello',
         time1: +new Date() + 86400000,
         time2: +new Date()
-    }, {
+    },
+    {
+        id: 8,
         name: 'ruby2',
         age: 15,
         dep: 3,
@@ -85,23 +102,20 @@ const tableData = [
         html: 'hello',
         time1: +new Date() + 86400000,
         time2: +new Date()
-    }
-];
+    }];
 
 new Vue({
     components: {
         'adc-table-column': column,
-        'adc-table': table
+        'adc-table': table,
+        pagination
     },
     el: '#app',
     data: {
         table: 'table1',
-        tableData: tableData,
+        tableData: [],
         loading: true,
-        selection: {
-            checkAll: false,
-            checked: [1, 2]
-        },
+        selection: [1, 2, 3, 4],
         visibleFilter(data) {
             return data > 13;
         },
@@ -109,10 +123,12 @@ new Vue({
             {
                 key: 1,
                 val: '系统研发'
-            }, {
+            },
+            {
                 key: 2,
                 val: '移动平台'
-            }, {
+            },
+            {
                 key: 3,
                 val: '移动发行'
             }
@@ -124,8 +140,8 @@ new Vue({
             img: {
                 name: 'adc-img',
                 config: {
-                    width: 180,
-                    height: 180
+                    width: 80,
+                    height: 80
                 }
             },
             operator: {
@@ -146,20 +162,22 @@ new Vue({
         },
         dateTimeFormat2(v) {
             return utility.dateFilter(v, 'yyyy-MM-dd HH:mm:ss');
-        }
+        },
+        pageIndex: 1
     },
-    created(){
-        setTimeout(()=>{
+    created() {
+        setTimeout(() => {
             this.loading = false;
+            this.tableData = tableData;
         }, 1000);
     },
     methods: {
-        clickSlot(data, item, index){
+        clickSlot(data, item, index) {
             console.log(data, item, index);
         },
         edit(item, index) {
             console.log(item, index);
-        },  
+        },
         getChecked() {
             console.log(this.$refs.table.getChecked());
         },
@@ -172,17 +190,22 @@ new Vue({
         changeTable() {
             this.table = 'table2';
         },
-        checkAll(checkAll){
+        checkAll(checkAll) {
             console.log(checkAll);
         },
-        checkOne(checked){
+        checkOne(checked) {
             console.log(checked);
         },
-        changeSelection(){
-            this.selection = {
-                checkAll: false,
-                checked: []
-            };
+        changeSelection() {
+            this.selection = [];
+        },
+        changePage(index){
+            this.pageIndex = index;
+            if (index === 1){
+                this.tableData = tableData;
+            } else {
+                this.tableData = tableData2;
+            }
         }
     }
 });