+++ src/js/module/common/buttons.vue
@@ -1,10 +1,13 @@
 <template>
     <div>
-        <button type="button" class="btn btn-sm btn-primary" @click="edit">
+        <button type="button"
+                class="btn btn-sm btn-primary"
+                @click="edit">
             <span class="glyphicon glyphicon-edit"></span>
             编辑
         </button>
-        <button type="button" class="btn btn-sm btn-success">
+        <button type="button"
+                class="btn btn-sm btn-success">
             <span class="glyphicon glyphicon-eye-open"></span>
             查看
         </button>
@@ -12,36 +15,36 @@
 </template>
 
 <script type="es6">
-    export default {
-        name: 'buttons',
-        props: {
-            prop: {
-                type: String,
-                default: ''
-            },
-            item: {
-                type: Object,
-                default() {
-                    return {};
-                }
-            },
-            index: {
-                type: Number,
-                default() {
-                    return -1;
-                }
-            },
-            data: {
-                type: Array,
-                default() {
-                    return [];
-                }
+export default {
+    name: 'buttons',
+    props: {
+        prop: {
+            type: String,
+            default: ''
+        },
+        item: {
+            type: Object,
+            default() {
+                return {};
             }
         },
-        methods: {
-            edit() {
-                console.log(this.item, this.index, this.data)
+        index: {
+            type: Number,
+            default() {
+                return -1;
+            }
+        },
+        data: {
+            type: Array,
+            default() {
+                return [];
             }
         }
-    };
+    },
+    methods: {
+        edit() {
+            console.log(this.item, this.index, this.data);
+        }
+    }
+};
 </script>